+++ packages/terra-content-container/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Updated to implement the new terra-scrollable-container component.
 
 2.11.0 - (May 30, 2018)
 ------------------